# src/main/resources/swagger.yaml
swagger: "2.0"
info:
  title: Practica 5
  version: '1.0.0'
  description: >
    Definicion de una API sencilla de reserva de salas para fines educativos.
    Cuenta con 5 tipos principales de objeto: Usuarios, Reservas, Edificios, Salas y Recursos.
  termsOfService: http://swagger.io/terms/
  contact:
    email: fermatfe@jcyl.es
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
...
host: localhost:8080
basePath: /v1
schemes:
- http
...
paths:
...
  /edificios:
    post:
      tags:
      - Edificios
      summary: Agregar edificio
      operationId: agregarEdificio
      description: Agrega un edificio al sistema
      consumes:
      - application/json
      parameters:
      - name: body
        in: body
        description: Datos del nuevo edificio
        required: true
        schema:
          $ref: '#/definitions/Edificio'
      responses:
        201:
          description: Edificio creado
        400:
          description: Peticion incorrecta
          schema:
            $ref: '#/definitions/Error'

    get:
      tags:
      - Edificios
      summary: Busca edificios
      operationId: getEdificios
      description: Busqueda de edificios
      produces:
      - application/json
      parameters:
      - name: searchString
        in: query
        description: Pasa una cadena opcional para buscar edificios
        required: false
        type: string
      - name: idProvincia
        in: query
        description: Busca los edificios de una provincia"
        required: false
        type: integer
        format: int32
      responses:
        200:
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/Edificio'
        404:
          description: No hay edificios con esos criterios
          schema:
            $ref: '#/definitions/Error'
        400:
          description: Peticion incorrecta
          schema:
            $ref: '#/definitions/Error'
  /edificios/{idEdificio}:
    get:
      tags:
      - Edificios
      summary: Busca edificio por ID
      description: Devuelve un edificio por ID
      operationId: getEdificioPorId
      produces:
      - application/json
      parameters:
      - name: idEdificio
        in: path
        description: ID del edificio a recuperar
        required: true
        type: integer
        format: int32
      responses:
        200:
          description: Edificio obtenido
          schema:
            $ref: '#/definitions/Edificio'
        404:
          description: Edificio no encontrado
          schema:
            $ref: '#/definitions/Error'
        400:
          description: Peticion incorrecta
          schema:
            $ref: '#/definitions/Error'

    put:
      tags:
      - Edificios
      summary: Actualizar edificio
      operationId: actualizarEdificio
      consumes:
      - application/json
      parameters:
        - name: idEdificio
          in: path
          description: ID del edificio a actualizar
          required: true
          type: integer
          format: int32
        - name: edificio
          in: body
          description: Nuevos Datos del edificio
          required: true
          schema:
            $ref: '#/definitions/Edificio'
      responses:
        200:
          description: OK
        400:
          description: ID invalido
          schema:
            $ref: '#/definitions/Error'
        404:
          description: Edificio no encontrado
          schema:
            $ref: '#/definitions/Error'
      
    delete:
      tags:
      - Edificios
      summary: Borra un edificio
      operationId: borrarEdificio
      parameters:
      - name: idEdificio
        in: path
        description: ID del edificio a borrar
        required: true
        type: integer
        format: int32
      responses:
        200:
          description: Edificio borrado
        404:
          description: Edificio no encontrado"
          schema:
            $ref: '#/definitions/Error'
        400:
          description: Peticion incorrecta
          schema:
            $ref: '#/definitions/Error'
  /edificios/{idEdificio}/salas:
    post:
      tags:
      - Edificios
      summary: Agregar sala
      operationId: agregarSala
      description: Agrega una sala al sistema
      consumes:
      - application/json
      parameters:
      - name: idEdificio
        in: path
        description: ID del edificio a recuperar
        required: true
        type: integer
        format: int32
      - name: body
        in: body
        description: Datos de la nueva sala
        schema:
          $ref: '#/definitions/Sala'
      responses:
        201:
          description: Sala creada
        400:
          description: "Peticion incorrecta"
          schema:
            $ref: '#/definitions/Error'

    get:
      tags:
      - Edificios
      summary: Buscar salas
      operationId: getSalas
      description: Busqueda de salas
      produces:
      - application/json
      parameters:
      - name: idEdificio
        in: path
        description: ID del edificio a recuperar
        required: true
        type: integer
        format: int32
      responses:
        200:
          description: Obtenidas las salas del edificio
          schema:
            type: array
            items:
              $ref: '#/definitions/Sala'
        400:
          description: Peticion incorrecta
          schema:
            $ref: '#/definitions/Error'
  /edificios/{idEdificio}/salas/{idSala}:  
    get:
      tags:
      - Edificios
      summary: Busca sala por ID
      description: Devuelve una sala por ID
      operationId: getSalaPorId
      produces:
      - application/json
      parameters:
      - name: idEdificio
        in: path
        description: ID del edificio a recuperar
        required: true
        type: integer
        format: int32
      - name: idSala
        in: path
        description: ID de la sala a recuperar
        required: true
        type: integer
        format: int32
      responses:
        200:
          description: Obtenida la sala del edificio
          schema:
            $ref: '#/definitions/Sala'
        404:
          description: Sala no encontrada
          schema:
            $ref: '#/definitions/Error'
        400:
          description: Peticion incorrecta
          schema:
            $ref: '#/definitions/Error'

    put:
      tags:
      - Edificios
      summary: Actualizar sala
      operationId: actualizarSala
      description: Actualiza una sala
      consumes:
      - application/json
      parameters:
        - name: idEdificio
          in: path
          description: ID del edificio a recuperar
          required: true
          type: integer
          format: int32
        - name: idSala
          in: path
          description: ID de la sala a actualizar
          required: true
          type: integer
          format: int32
        - name: sala
          in: body
          description: Nuevos Datos de la sala
          required: true
          schema:
            $ref: '#/definitions/Sala'
      responses:
        200:
          description: OK
        400:
          description: "Peticion incorrecta"
          schema:
            $ref: '#/definitions/Error'
        404:
          description: "Sala no encontrada"
          schema:
            $ref: '#/definitions/Error'

    delete:
      tags:
      - Edificios
      summary: Borra una sala
      operationId: borrarSala
      parameters:
      - name: idEdificio
        in: path
        description: ID del edificio a recuperar
        required: true
        type: integer
        format: int32
      - name: idSala
        in: path
        description: ID de la sala a borrar
        required: true
        type: integer
        format: int32
      responses:
        200:
          description: Sala borrada
        404:
          description: Sala no encontrada
          schema:
            $ref: '#/definitions/Error'
        400:
          description: Peticion incorrecta
          schema:
            $ref: '#/definitions/Error'
definitions:
  Provincia:
    type: object
    required:
    - id
    - provincia
    properties:
      id:
        type: integer
        description: Identificador de la provincia
        example: 40
      provincia:
        type: string
        description: Nombre de la provincia
        example: Valladolid
  Poblacion:
    type: object
    required:
    - id
    - poblacion
    properties:
      id:
        type: integer
        description: Identificador de la población
        example: 7467
      poblacion:
        type: string
        description: Nombre de la Población
        example: Valladolid"
      latitud:
        type: string
        description: Latitud de la población (geoposicionamiento)
        example: "41.653363"
      longitud:
        type: string
        description: Longitud de la población (geoposicionamiento)"
        example: "-4.728874"
      provincia:
        $ref: '#/definitions/Provincia'
  Sala:
    type: object
    required:
    - nombre
    properties:
      id:
        type: integer
        description: Identificador de la Sala
        example: 1
      nombre:
        type: string
        description: Nombre de la sala
        example: Sala XXX
      descripcion:
        type: string
        example: Sala de reuniones nº1
      capacidad:
        type: integer
        example: 10
      localizacion:
        type: string
        description: Localización de la sala
        example: "SM-24"
      tipoSala:
        $ref: '#/definitions/TipoSala'
      edificio:
         $ref: '#/definitions/Edificio'
  TipoSala:
    type: object
    required:
    - id
    - tipo
    properties:
      id:
        type: integer
        description: "Id de Tipo de Sala"
        example: 1
      tipo:
        type: string
        description: "Nombre de Tipo de Sala"
        example: "Sala de reunion"
...
  Edificio:
    type: object
    required:
    - nombre    
    properties:
      id:
        type: integer
        description: Identificador del edificio
        example: 1
      nombre:
        type: string
        description: Nombre del Edificio
        example: E.U.M. I
      direccion:
        $ref: '#/definitions/Direccion'
      titularidad:
        type: string
        description: Titularidad del edificio
        example: 'PR'
  Direccion:
    type: object
    required:
    - nombreVia
    - numeroVia
    - codigoPostal
    properties:
      tipoVia:
        type: string
        description: Codigo de via
        example: Calle
      nombreVia:
        type: string
        description: Nombre de via
        example: Duque de la Victoria
      numeroVia:
        type: string
        description: Numero de via
        example: "20"
      codigoPostal:
        type: string
        description: Codigo postal del Edificio
        example: "47002"
      poblacion:
        $ref: '#/definitions/Poblacion'
  Error:
    type: object
    properties:
      timestamp:
        type: string
        format: date-time
      status:
        type: integer
        format: int32
      error:
        type: string
      message:
        type: string
      path:
        type: string
    example:
      timestamp: '2018-06-08T17:04:22.893+0000'
      status: 404
      error: Not Found
      message: No encontrado el recurso
      path: v1/provincias